refactor: delegate utils
