+++ programs/svm-spoke/src/utils/mod.rs
@@ -1,7 +1,7 @@
 pub mod bitmap_utils;
 pub mod cctp_utils;
+pub mod delegate_utils;
 pub mod deposit_utils;
-pub mod fill_utils;
 pub mod merkle_proof_utils;
 pub mod message_utils;
 pub mod testable_utils;
@@ -9,8 +9,8 @@ pub mod transfer_utils;
 
 pub use bitmap_utils::*;
 pub use cctp_utils::*;
+pub use delegate_utils::*;
 pub use deposit_utils::*;
-pub use fill_utils::*;
 pub use merkle_proof_utils::*;
 pub use message_utils::*;
 pub use testable_utils::*;
